pyproject.toml
[project]
name = "lottery-tg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "python-telegram-bot>=22.1",
]
